added --no-auto-delete option

+++ sansyo_count.cpp
@@ -1,5 +1,7 @@
 #include "sansyo_count.h"
 
+bool sansyo_all_t::do_auto_delete=true;
+
 template<>
 bool sansyo_t<kankyo_t>::check_sansyo() const {
 	return sansyo==NULL || creater_t::creater().is_exist_kankyo(sansyo);
@@ -14,7 +16,7 @@ template<>
 void sansyo_t<kankyo_t>::decrement_sansyo_count() {
 	if(sansyo!=NULL) {
 		sansyo->sansyo_count--;
-		if(sansyo->sansyo_count<=0) {
+		if(do_auto_delete && sansyo->sansyo_count<=0) {
 			creater_t::creater().delete_kankyo(sansyo);
 			sansyo=NULL;
 		}
@@ -25,7 +27,7 @@ template<>
 void sansyo_t<data_t>::decrement_sansyo_count() {
 	if(sansyo!=NULL) {
 		sansyo->sansyo_count--;
-		if(sansyo->sansyo_count<=0) {
+		if(do_auto_delete && sansyo->sansyo_count<=0) {
 			creater_t::creater().delete_data(sansyo);
 			sansyo=NULL;
 		}
